requirements.txt
schedule==1.2.1
tweepy==4.14.0
python-dotenv==1.0.1